requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
dj-rest-auth==1.1.0
Django==3.1
django-allauth==0.42.0
djangorestframework==3.11.2
idna==2.10
oauthlib==3.1.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.6
